Friend Page
I called it notification page before I realised it should be called friend page.

This code accesses a list I added to the route. So if there is nothing in the list i will show a message saying that the user has no friend requests. Same for friends (no friends in list = no friends message on page). Hopefully this is easy enough to integrate into database. Lmk if not :D